[FIX] fix link error and readme error



# message auto-generated for no-merge-commit merge:
!12 merge master into master

[FIX] fix link error and readme error

Created-by: chenhongyang
Commit-by: gcw_roGn03Sa
Merged-by: turing_project1
Description: # Pull Request
## 描述
增加对文档及环境部署上的描述更新

## 类型
[x] Bug修复
[ ] 新功能
[ ] 重构（即不是新增功能，也不是修改bug的代码变动）
[ ] 构建过程或辅助工具的变动
[x] 文档内容更新

如何测试
```
source /usr/local/Ascend/ascend-toolkit/set_env.sh
python demo_infer.py --ckpt ckpt/model.pt
```

## checklist
[x] 我的代码遵循这个项目的代码风格
[x] 我已经自己测试过我的代码
[x] 我已经更新了相应的文档
[x] 我已经在标题中正确使用了类型标签（例如：feat, fix, refactor, docs, test）

See merge request: cann/cann-recipes-spatial-intelligence!12

+++ models/vggt/eval/point_map_estimation/utils.py
@@ -122,7 +122,7 @@ def extract_pts3d(criterion, data, pred_result, use_proj):
             pts = point_map_by_unprojection[j]
             conf = depth_conf[0, j].to(torch.float).cpu().data.numpy()
         else:
-            pts = pred_pts[j].cpu().numpy()[0]
+            pts = pred_pts[j].to(torch.float).cpu().numpy()[0]
             conf = preds[j]["conf"].to(torch.float).cpu().data.numpy()[0]
 
         pts_gt = gt_pts[j].detach().to(torch.float).cpu().numpy()[0]
